Refactor app to use svelte stores
After a LOT of faffing around trying to find a router I thought I needed
I ended up sticking with the original choice (with hashes) because it
did the actual routing part. I got hung up on trying to pass data to the
child components via the router.

The advice on one of the issues was to use Svelte stores. So I have done
this.

In addition to this I have also renamed some things to match what they
are and set up the project for the reskin